Move memkind_memtier header to internal API
