+++ tiering/ctl.c
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: BSD-2-Clause
 /* Copyright (C) 2021 Intel Corporation. */
 
-#include <memkind_memtier.h>
+#include <memkind/internal/memkind_memtier.h>
 #include <tiering/memtier_log.h>
 
 #include <errno.h>
