+++ tiering/memtier.c
@@ -2,7 +2,7 @@
 /* Copyright (C) 2021 Intel Corporation. */
 
 #include "../config.h"
-#include <memkind_memtier.h>
+#include <memkind/internal/memkind_memtier.h>
 
 #include <tiering/ctl.h>
 #include <tiering/memtier_log.h>
